// src/app/admin/employees/type.ts
export type EmployeesVo = {
  name: string;
  workerRoleIds: (string | undefined)[] | undefined;
  firstName: string;
  id: string;
  lastName: string;
  middleName: string;
  externalId: string;
  email: string;
  phone: string;
  license: string;
  credential: string;
  hireDate: string;
  targetedHoursPerWeek: string;
  terminationDate: string;
  departmentIds: Array<string>;
  departments?: { deptId: string; name: string }[];
  status: number;
  userId: string;
  workerId: string;
  autoInvite?: boolean;
  roleIds?: string[];
  roles: { id: string }[];
  userCommunityId: string;
  userCommunityRefId: string;
  checkInTime?: string;
  checkOutTime?: string;
  exceptions: string;
  nationalPhone: string;
};